LayoutTests/http/tests/security/xssAuditor/script-tag-with-invalid-url-encoding-expected.txt

CONSOLE MESSAGE: line 1: Refused to execute a JavaScript script. Source code of script found within request.
